requirements.txt

flask
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv